renames BuildIDComparison since we use it for versions as well
