+++ src/app/patterns/body/explorer/CompareOverTimeGraph.svelte
