+++ src/app/patterns/body/explorer/QuantileExplorer.svelte
@@ -5,7 +5,7 @@ import { tweened } from 'svelte/motion';
 import { cubicOut as easing } from 'svelte/easing';
 import { format } from 'd3-format';
 
-import BuildIDComparison from './BuildIDComparison.svelte';
+import CompareOverTimeGraph from './CompareOverTimeGraph.svelte';
 // import TotalClientsGraph from './TotalClientsGraph.svelte';
 import DistributionComparison from './DistributionComparison.svelte';
 import ComparisonSummary from './ComparisonSummary.svelte';
@@ -110,6 +110,11 @@ $: if (reference.histogram) animatedReferenceDistribution.setValue(reference.his
   grid-template-columns: max-content max-content auto;
 }
 
+.no-line-chart {
+  grid-template-columns: max-content auto;
+  justify-items: start;
+}
+
 .summary {
   display:grid;
   grid-auto-flow:column;
@@ -163,9 +168,9 @@ h4 {
   </div>
 </div>
 
-<div class=graphic-and-summary>
+<div class=graphic-and-summary class:no-line-chart={insufficientData}>
     <div style="display: {insufficientData ? 'none' : 'block'}">
-      <BuildIDComparison
+      <CompareOverTimeGraph
         data={data}
         xDomain={$domain}
         yDomain={yDomain}
